requirements.txt
# Web framework
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
jinja2>=3.1.0
python-multipart>=0.0.9

# Database
sqlalchemy>=2.0.0
aiosqlite>=0.19.0

# Podcast/YouTube downloading
yt-dlp>=2025.1.0

# Transcription
faster-whisper>=1.0.0

# HTTP requests (for SponsorBlock API)
httpx>=0.27.0

# RSS feed generation
feedgen>=1.0.0

# Scheduling
apscheduler>=3.10.0

# Utilities
pydantic>=2.6.0
pydantic-settings>=2.1.0
python-slugify>=8.0.0
